gopy: add support for sub-tests
This CL adds support for sub-tests for Go >= 1.7 and adds the needed
fallback code for Go < 1.7.

This allows to run sub-tests in parallel, where each sub-test tests a
backend (py2-cffi, py3-cffi, py3, ...).
